Update index.js
The mysql-events initialization method has been changed to support custom error handling in the file where this module is used.
The previous way of handling did not work for some specific situations like recovering when the database is restarted. The new connection was set up and was active on node process level, but all connection objects of the event were indicated as disconnected and log updates or event triggers were not detected. Sometimes a developer should decide just to create completely new MySQLEvents object and therefore the custom event handling should be implemented for more flexibility.